requirements.txt
streamlit>=1.40.0
google-genai>=1.56.0
python-dotenv>=1.0.1